+++ frontend/src/components/Admin/AdminCategory.vue
@@ -56,7 +56,6 @@ export default {
 	name: 'AdminCategory',
 	data() {
 		return {
-			user_idx: this.$session.get('user_idx'),
 			categoryList: null,
 			edit: false,
 			category: {
@@ -126,9 +125,9 @@ export default {
 		},
 	},
 	created() {
-		if (this.user_idx !== 1) {
+		if (this.$session.get('rating') !== '관리자') {
 			alert('잘못된 접근방법입니다');
-			this.$router.go(-1);
+			this.$router.push('/');
 		} else {
 			this.$http.get(`/api/admin/categoryList`).then(res => {
 				this.categoryList = res.data;