+++ frontend/src/components/Admin/AdminUser.vue
@@ -55,7 +55,6 @@ export default {
 	name: 'AdminUser',
 	data() {
 		return {
-			user_idx: this.$session.get('user_idx'),
 			userList: null,
 			edit: false,
 			user: {
@@ -112,9 +111,9 @@ export default {
 		},
 	},
 	created() {
-		if (this.user_idx !== 1) {
+		if (this.$session.get('rating') !== '관리자') {
 			alert('잘못된 접근방법입니다');
-			this.$router.go(-1);
+			this.$router.push('/');
 		} else {
 			this.$http.get(`/api/admin/userList`).then(res => {
 				this.userList = res.data;