+++ frontend/src/components/Board/CategoryList.vue
@@ -16,7 +16,7 @@
 				</thead>
 				<tbody>
 					<tr v-if="boardList === null">
-						<td colspan="3" style="text-align: center;">데이터가 없습니다</td>
+						<td colspan="4" style="text-align: center;">데이터가 없습니다</td>
 					</tr>
 					<tr v-else v-for="(board, bIdx) in displayedPosts" :key="bIdx" @click="detail(board.post_id)">
 						<th scope="row">{{ board.post_id }}</th>
@@ -48,7 +48,9 @@
 							</button>
 						</li>
 						<li class="page-item">
-							<button type="button" @click="pageNum++" v-if="pageNum < totalPage.length" class="page-link">다음</button>
+							<button type="button" @click="pageNum++" v-if="pageNum < totalPage.length" class="page-link">
+								다음
+							</button>
 						</li>
 					</ul>
 				</nav>
@@ -163,7 +165,7 @@ export default {
 
 <style>
 #pageStyle {
-	float: left;
+	margin: 0 auto;
 }
 
 button.page-link {