+++ frontend/src/main.js
@@ -6,6 +6,7 @@ import BootstrapVue from 'bootstrap-vue';
 import moment from 'moment';
 import VueMomentJS from 'vue-momentjs';
 import VueSession from 'vue-session';
+import jQuery from 'jquery';
 
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'bootstrap-vue/dist/bootstrap-vue.css';
@@ -16,6 +17,7 @@ Vue.use(VueMomentJS, moment);
 
 Vue.config.productionTip = false;
 Vue.prototype.$http = http;
+Vue.prototype._$ = jQuery;
 
 new Vue({
 	router,