+++ frontend/src/views/Home.vue
@@ -26,10 +26,11 @@
 					<b-button @click="login" class="loginbutton" align="center">Login</b-button>
 				</div>
 				<div v-else>
-					{{ this.$session.get('rating') }}
-					{{ this.$session.get('name') }}님 환영 합니다!
-					<button v-on:click="Mypage">마이페이지</button>
-					<button v-on:click="logout">로그아웃</button>
+					<img alt="관리자" width="16" height="16" src="/css/rating.png" />
+					[{{ this.$session.get('rating') }}]<br />
+					{{ this.$session.get('name') }}님 환영 합니다! <br />
+					<b-button variant="primary" v-on:click="Mypage">마이페이지</b-button>&nbsp;
+					<b-button variant="danger" v-on:click="logout">로그아웃</b-button>
 				</div>
 
 				<!-- TODO: 리스트 CSS 수정 -->